refactor: replace create-project-modal into project domain
