+++ libs/domains/projects/feature/src/index.ts
@@ -10,3 +10,4 @@ export * from './lib/hooks/use-edit-deployment-rule/use-edit-deployment-rule'
 export * from './lib/hooks/use-delete-deployment-rule/use-delete-deployment-rule'
 export * from './lib/hooks/use-edit-deployment-rules-priority-order/use-edit-deployment-rules-priority-order'
 export * from './lib/project-avatar/project-avatar'
+export * from './lib/create-project-modal/create-project-modal'
