+++ libs/shared/console-shared/src/index.ts
@@ -16,7 +16,6 @@ export * from './lib/flow-create-port/ui/crud-modal/crud-modal'
 export * from './lib/flow-create-port/utils/port-healthcheck'
 export * from './lib/git-repository-settings/ui/git-repository-settings/git-repository-settings'
 export * from './lib/git-repository-settings/feature/edit-git-repository-settings-feature/edit-git-repository-settings-feature'
-export * from './lib/create-project-modal/feature/create-project-modal-feature'
 export * from './lib/member-invitation/accept-invitation/feature/accept-invitation-feature'
 export * from './lib/cluster-settings/ui/cluster-general-settings/cluster-general-settings'
 export * from './lib/cluster-settings/feature/cluster-credentials-settings-feature/cluster-credentials-settings-feature'
